Refactor: Flatten project structure
- Moved all Swift code from AppDelegate.swift into main.swift.
- Deleted AppDelegate.swift.
- Moved Assets.xcassets and ChronoMouse.entitlements to the root directory.
- Removed the ChronoMouse subdirectory.
- Updated Xcode project file to reflect these changes.
