+++ Assets.xcassets/AppIcon.appiconset/ChronoMouse256.png
