+++ Assets.xcassets/Contents.json
